// src/app/sidur_avoda/shared/types/pirtei_oved.ts
export interface IPirteiOved {
    shem_prati?: string;
    shem_mishpacha?: string;
    td_zehut_user?: string;
    k_user?: number;
    mispar_nahag?: number;
    full_name?: string;
    id_pirtei_nahag?: number;
    mispar_oved?: string;
    yechida?: string;
    t_tafkid?: string;
    tr_end?: Date;
    id_maamad?: number;
    k_maamd?: string;
    teur_maamd?: string;
    sw_payil_lo_payil?: number;
    shabat_chag?: number;
    id_yechida?: number;
    kod_yechida?: string;
    global_tafkid?: number;
    k_tafkid_beagaf_piku?: number;
    k_tafkid?: number;
    mispar_rishayon?: string;
    tokef_rishayon?: Date;
    id_hagbalat_rishayon?: number;
    teur_hagbalat_rishayon?: string;
    me_taarich?: Date;
    ad_taarich?: Date;
    hearat_rishayon?: string;
    ms_mefakeach_pikuach?: number;
    shem_mishtamesh_maar?: string;
    ms_brirat_knas?: number;
    sugeiRechev?: string;
    sugeiPeilut?: string;
    tr_start?: Date;
    mesimotMeshubetzot?: string;
    mesimotTemplate?: string;
    teurHeadrut?: string;
    ovedNiud?: string;
    hagbalatRishayon?: string;
    heara?: string;
    deleted?: number;
    sug_record?: number;
}